% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posteriorPLSR.R
\name{posteriorPLSRReconstruction}
\alias{posteriorPLSRReconstruction}
\title{posteriorPLSRReconstruction}
\usage{
posteriorPLSRReconstruction(exprData, adjMatrix, type = "expr",
  components = 3, timepoints = 8, iterations = 1000, mu = 1,
  sigmaX = 0.1, sigmaW = 0.1, sigmaE = 0.1, p = 0.5, q = 0.5,
  cores = 1)
}
\arguments{
\item{exprData}{expression matrix with rows corresponding to timepoints in increasing temporal order and columns to genes; or a list containing two matrices \code{xData} (inputs) and \code{yData} (responses) with rows corresponding to samples and columns to genes}

\item{adjMatrix}{adjacency matrix}

\item{type}{gene expression model to use; 'expr' or 'deltaExpr'}

\item{components}{number of PLSR components to use}

\item{timepoints}{number of timepoints to use in the generated expression data}

\item{iterations}{number of VIP scores to generate}

\item{mu}{mean edge weight}

\item{sigmaX}{edge weight standard deviation}

\item{sigmaW}{initial condition standard deviation}

\item{sigmaE}{generated data noise}

\item{p}{probability of an edge, given that the pair was a prior edge}

\item{q}{probability of an edge, given that the pair was a prior non-edge}

\item{cores}{number of cores for parallelization}
}
\value{
a data frame containing the data-derived VIP scores, estimated number of null VIP scores that are less than the data-derived score in the edge and non-edge cases, and (optionally) posterior edge probabilities for each ordered pair of genes.
}
\description{
Computes null distrbution-related results and computes posterior edge probabilities using a prior network and input expression data
}
